feat: Add review comments functionality to employee form
- Updated employee dialog to display employee ID instead of national ID.
- Added a textarea for review comments in the employee form.
- Modified the form initialization to include review comments.
- Enhanced form submission logic to handle review comments.
- Updated form details component to conditionally display review comments.
- Added styling for reviewed and not reviewed states in the table.
- Adjusted API models and services to support review comments.
- Changed data types for amount in reports from double to decimal for accuracy.
- Updated API content paths for images and fonts.

+++ src/Application/Features/DailyService.cs
@@ -282,7 +282,7 @@ namespace Application.Features
             dt2.Columns.Add("كود تجارة", typeof(string));
             dt2.Columns.Add("القسم", typeof(string));
             dt2.Columns.Add("الاسم", typeof(string));
-            dt2.Columns.Add("المبلغ", typeof(double));
+            dt2.Columns.Add("المبلغ", typeof(decimal));
             int counter2 = 1;
             foreach (var item in dailyToDataTable)
             {
@@ -293,7 +293,7 @@ namespace Application.Features
                 dr["كود تجارة"] = item.TegaraCode;
                 dr["القسم"] = item.Department == null ? "" : item.Department.Name;
                 dr["الاسم"] = item.Name;
-                dr.SetField("المبلغ", (double)item.Amount);
+                dr.SetField("المبلغ", (decimal)item.Amount);
                 dt2.Rows.Add(dr);
 
             }
